feat: change Rating field type from int to float64 in UserBookProgress and UserBookProgressGORM models

+++ internal/domain/book_progress.go
@@ -23,7 +23,7 @@ type UserBookProgress struct {
 	HighestScore         float64
 	LastPageID           uint
 	CurrentSessionPoints int
-	Rating               int
+	Rating               float64
 	CreatedAt            time.Time
 	UpdatedAt            time.Time
 	DeletedAt            gorm.DeletedAt
